// client/src/app/service/api-client.ts
export const API_BASE_URL = new InjectionToken<string>('API_BASE_URL');

@Injectable({
    providedIn: 'root'
})
export class CpuClient {
    private http: HttpClient;
    private baseUrl: string;
    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;

    constructor(@Inject(HttpClient) http: HttpClient, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
        this.http = http;
        this.baseUrl = baseUrl ?? "";
    }

    getCpu(name_Operator?: ConditionType | null | undefined, name_Values?: string[] | undefined, sortProperty?: string | null | undefined, sortDirection?: SortType | null | undefined, pageNumber?: number | null | undefined, pageSize?: number | null | undefined): Observable<PagedResponseOfListOfCpuDTO> {
        let url_ = this.baseUrl + "/api/cpu?";
        if (name_Operator !== undefined && name_Operator !== null)
            url_ += "Name.Operator=" + encodeURIComponent("" + name_Operator) + "&";
        if (name_Values === null)
            throw new globalThis.Error("The parameter 'name_Values' cannot be null.");
        else if (name_Values !== undefined)
            name_Values && name_Values.forEach(item => { url_ += "Name.Values=" + encodeURIComponent("" + item) + "&"; });
        if (sortProperty !== undefined && sortProperty !== null)
            url_ += "SortProperty=" + encodeURIComponent("" + sortProperty) + "&";
        if (sortDirection !== undefined && sortDirection !== null)
            url_ += "SortDirection=" + encodeURIComponent("" + sortDirection) + "&";
        if (pageNumber !== undefined && pageNumber !== null)
            url_ += "PageNumber=" + encodeURIComponent("" + pageNumber) + "&";
        if (pageSize !== undefined && pageSize !== null)
            url_ += "PageSize=" + encodeURIComponent("" + pageSize) + "&";
        url_ = url_.replace(/[?&]$/, "");

        let options_ : any = {
            observe: "response",
            responseType: "blob",
            headers: new HttpHeaders({
                "Accept": "application/json"
            })
        };

        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
            return this.processGetCpu(response_);
        })).pipe(_observableCatch((response_: any) => {
            if (response_ instanceof HttpResponseBase) {
                try {
                    return this.processGetCpu(response_ as any);
                } catch (e) {
                    return _observableThrow(e) as any as Observable<PagedResponseOfListOfCpuDTO>;
                }
            } else
                return _observableThrow(response_) as any as Observable<PagedResponseOfListOfCpuDTO>;
        }));
    }

    protected processGetCpu(response: HttpResponseBase): Observable<PagedResponseOfListOfCpuDTO> {
        const status = response.status;
        const responseBlob =
            response instanceof HttpResponse ? response.body :
            (response as any).error instanceof Blob ? (response as any).error : undefined;

        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
        if (status === 200) {
            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
            let result200: any = null;
            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
            result200 = PagedResponseOfListOfCpuDTO.fromJS(resultData200);
            return _observableOf(result200);
            }));
        } else if (status !== 200 && status !== 204) {
            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            }));
        }
        return _observableOf(null as any);
    }
}
// ...
export class PagedResponseOfListOfCpuDTO implements IPagedResponseOfListOfCpuDTO {
    data?: CpuDTO[] | undefined;
    pageNumber?: number;
    pageSize?: number;
    totalRecords?: number;

    constructor(data?: IPagedResponseOfListOfCpuDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["data"])) {
                this.data = [] as any;
                for (let item of _data["data"])
                    this.data!.push(CpuDTO.fromJS(item));
            }
            this.pageNumber = _data["pageNumber"];
            this.pageSize = _data["pageSize"];
            this.totalRecords = _data["totalRecords"];
        }
    }

    static fromJS(data: any): PagedResponseOfListOfCpuDTO {
        data = typeof data === 'object' ? data : {};
        let result = new PagedResponseOfListOfCpuDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.data)) {
            data["data"] = [];
            for (let item of this.data)
                data["data"].push(item ? item.toJSON() : undefined as any);
        }
        data["pageNumber"] = this.pageNumber;
        data["pageSize"] = this.pageSize;
        data["totalRecords"] = this.totalRecords;
        return data;
    }
}

export interface IPagedResponseOfListOfCpuDTO {
    data?: CpuDTO[] | undefined;
    pageNumber?: number;
    pageSize?: number;
    totalRecords?: number;
}

export class BaseEntity implements IBaseEntity {
    id?: number;

    constructor(data?: IBaseEntity) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
        }
    }

    static fromJS(data: any): BaseEntity {
        data = typeof data === 'object' ? data : {};
        let result = new BaseEntity();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        return data;
    }
}

export interface IBaseEntity {
    id?: number;
}

export class CpuDTO extends BaseEntity implements ICpuDTO {
    name?: string;
    cores?: number;
    threads?: number;
    baseClock?: number;
    boostClock?: number;
    tdp?: number;
    socket?: string;
    releaseYear?: number;
    price?: number;
    description?: string;
    manufacturerDboId?: number;
    manufacturer?: ManufacturerDTO;

    constructor(data?: ICpuDTO) {
        super(data);
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.name = _data["name"];
            this.cores = _data["cores"];
            this.threads = _data["threads"];
            this.baseClock = _data["baseClock"];
            this.boostClock = _data["boostClock"];
            this.tdp = _data["tdp"];
            this.socket = _data["socket"];
            this.releaseYear = _data["releaseYear"];
            this.price = _data["price"];
            this.description = _data["description"];
            this.manufacturerDboId = _data["manufacturerDboId"];
            this.manufacturer = _data["manufacturer"] ? ManufacturerDTO.fromJS(_data["manufacturer"]) : undefined as any;
        }
    }

    static override fromJS(data: any): CpuDTO {
        data = typeof data === 'object' ? data : {};
        let result = new CpuDTO();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["name"] = this.name;
        data["cores"] = this.cores;
        data["threads"] = this.threads;
        data["baseClock"] = this.baseClock;
        data["boostClock"] = this.boostClock;
        data["tdp"] = this.tdp;
        data["socket"] = this.socket;
        data["releaseYear"] = this.releaseYear;
        data["price"] = this.price;
        data["description"] = this.description;
        data["manufacturerDboId"] = this.manufacturerDboId;
        data["manufacturer"] = this.manufacturer ? this.manufacturer.toJSON() : undefined as any;
        super.toJSON(data);
        return data;
    }
}

export interface ICpuDTO extends IBaseEntity {
    name?: string;
    cores?: number;
    threads?: number;
    baseClock?: number;
    boostClock?: number;
    tdp?: number;
    socket?: string;
    releaseYear?: number;
    price?: number;
    description?: string;
    manufacturerDboId?: number;
    manufacturer?: ManufacturerDTO;
}

export class ManufacturerDTO implements IManufacturerDTO {
    id?: number;
    name?: string;
    description?: string;

    constructor(data?: IManufacturerDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): ManufacturerDTO {
        data = typeof data === 'object' ? data : {};
        let result = new ManufacturerDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        return data;
    }
}

export interface IManufacturerDTO {
    id?: number;
    name?: string;
    description?: string;
}

export enum ConditionType {
    Equals = "Equals",
    GreaterThan = "GreaterThan",
    LessThan = "LessThan",
    GreaterThanOrEqual = "GreaterThanOrEqual",
    LessThanOrEqual = "LessThanOrEqual",
    Between = "Between",
    StartsWith = "StartsWith",
    Contains = "Contains",
    In = "In",
    NotIn = "NotIn",
    NotEquals = "NotEquals",
    EndsWith = "EndsWith",
    IsNull = "IsNull",
    IsNotNull = "IsNotNull",
}

export enum SortType {
    Ascending = "Ascending",
    Descending = "Descending",
}
// ...
export class SwaggerException extends Error {
    override message: string;
    status: number;
    response: string;
    headers: { [key: string]: any; };
    result: any;

    constructor(message: string, status: number, response: string, headers: { [key: string]: any; }, result: any) {
        super();

        this.message = message;
        this.status = status;
        this.response = response;
        this.headers = headers;
        this.result = result;
    }

    protected isSwaggerException = true;

    static isSwaggerException(obj: any): obj is SwaggerException {
        return obj.isSwaggerException === true;
    }
}